+++ crates/components-runtime/src/context.rs
@@ -8,9 +8,6 @@ use wasmtime_wasi::{IoView, ResourceTable, WasiCtx, WasiView};
 
 use crate::config::{ComponentsConfiguration, DataCollectionComponents};
 use crate::consent_mapping::{ConsentMapping, ConsentMappingPre};
-use crate::data_collection::v0_5_0::data_collection::{
-    DataCollectionZeroFiveZero, DataCollectionZeroFiveZeroPre,
-};
 use crate::data_collection::v1_0_0::data_collection::{
     DataCollectionOneZeroZero, DataCollectionOneZeroZeroPre,
 };
@@ -22,7 +19,6 @@ pub struct ComponentsContext {
 }
 
 pub struct Components {
-    pub data_collection_0_5_0: HashMap<String, DataCollectionZeroFiveZeroPre<HostState>>,
     pub data_collection: HashMap<String, DataCollectionOneZeroZeroPre<HostState>>,
     pub consent_mapping: HashMap<String, ConsentMappingPre<HostState>>,
 }
@@ -42,27 +38,6 @@ pub fn pre_instanciate_data_collection_component_internal(
     let component = Component::from_file(engine, &component_config.file)?;
     let instance_pre = linker.instantiate_pre(&component)?;
     let instance_pre = DataCollectionOneZeroZeroPre::new(instance_pre)?;
-
-    tracing::debug!("loaded new data collection component");
-
-    Ok(instance_pre)
-}
-
-pub fn pre_instanciate_data_collection_0_5_0_component_internal(
-    engine: &Engine,
-    component_config: &DataCollectionComponents,
-) -> anyhow::Result<DataCollectionZeroFiveZeroPre<HostState>> {
-    let mut linker = Linker::new(engine);
-    wasmtime_wasi::add_to_linker_async(&mut linker)?;
-
-    let span = tracing::info_span!("component-context", component = %component_config.id, category = "data-collection");
-    let _span = span.enter();
-
-    tracing::debug!("Loading new data collection component");
-
-    let component = Component::from_file(engine, &component_config.file)?;
-    let instance_pre = linker.instantiate_pre(&component)?;
-    let instance_pre = DataCollectionZeroFiveZeroPre::new(instance_pre)?;
 
     tracing::debug!("loaded new data collection component");
 
@@ -98,17 +73,6 @@ impl ComponentsContext {
             })
             .collect::<anyhow::Result<_>>()?;
 
-        let data_collection_0_5_0_components = config
-            .data_collection
-            .iter()
-            .filter(|entry| entry.wit_version == DataCollectionWitVersion::V0_5_0)
-            .map(|entry| {
-                let instance_pre =
-                    pre_instanciate_data_collection_0_5_0_component_internal(&engine, entry)?;
-                Ok((entry.id.clone(), instance_pre))
-            })
-            .collect::<anyhow::Result<_>>()?;
-
         // Consent mapping components
         let consent_mapping_components = config
             .consent_mapping
@@ -131,7 +95,6 @@ impl ComponentsContext {
 
         let components = Components {
             data_collection: data_collection_components,
-            data_collection_0_5_0: data_collection_0_5_0_components,
             consent_mapping: consent_mapping_components,
         };
 
@@ -177,20 +140,6 @@ impl ComponentsContext {
         store: &mut Store<HostState>,
     ) -> anyhow::Result<DataCollectionOneZeroZero> {
         let instance_pre = self.components.data_collection.get(id);
-
-        if instance_pre.is_none() {
-            return Err(anyhow::anyhow!("component not found: {}", id));
-        }
-
-        instance_pre.unwrap().instantiate_async(store).await
-    }
-
-    pub async fn get_data_collection_0_5_0_instance(
-        &self,
-        id: &str,
-        store: &mut Store<HostState>,
-    ) -> anyhow::Result<DataCollectionZeroFiveZero> {
-        let instance_pre = self.components.data_collection_0_5_0.get(id);
 
         if instance_pre.is_none() {
             return Err(anyhow::anyhow!("component not found: {}", id));
