+++ crates/components-runtime/src/data_collection/v1_0_0/mod.rs
@@ -1,4 +1,5 @@
 mod convert;
+pub mod execute;
 
 pub mod data_collection {
     wasmtime::component::bindgen!({
